refactor(types): activate PositionBufferDescriptor — fix rawClustersBuffer type bug + remove as any
Follow-up to f952fbec + 81c92edc. The dead PositionBufferDescriptor type
documented a real contract that was being violated by a type bug, masked
by an as any cast. This activates the type by closing the contract.

The bug
-------
appState.rawClustersBuffer was typed as Float32Array | null in 3 places,
but the data loader produces Uint16Array (per BusinessDataResult and
PositionBufferDescriptor). The producer-consumer mismatch was papered
over by  at
data-store.ts:276 (now 293).

3 type sites carried the wrong annotation (all synced from the same source):
  - src/lib/state/app.svelte.ts:114          (the  field)
  - src/lib/state/state-types.ts:469         (SemanticState interface)
  - src/lib/engine/webgl-context.ts:33       (WebGLContextState interface)

The fix
-------
1. All 3 sites: Float32Array | null -> Uint16Array | null
2. data-store.ts: remove the  cast (types now align)
3. data-store.ts: re-add PositionBufferDescriptor import + add the
   canonical accessor function that the type was always meant to enable:
     export function getPositionBufferDescriptor(): PositionBufferDescriptor | null
   Bundles positionBuffer + clustersBuffer + count into the documented
   shape. Any future engine refactor that wants {buffer, count, clusters}
   imports this instead of re-composing the two writables.

Verification
------------
- typecheck: 0 new errors (the 11+ errors in arrival-handoff.ts,
  thread-inspector*.ts, thread-settler.ts are pre-existing dirty work
  from parallel sessions, confirmed by stash-and-rerun baseline test)
- vitest: 22/22 pass on data-store + as-any-budget test
- as any budget: 478 -> 477 (typing contract test now allows the lower
  count, locking in the win)
- vite build: clean in 17.54s

Diff: +36 / -23 across 4 files.

+++ src/lib/engine/webgl-context.ts
@@ -30,7 +30,7 @@ export interface WebGLContextState {
   nodeSporeHitMesh: InstancedMesh | null;
   nodeSporeMaterial: Material | null;
   rawPositionsBuffer: Float32Array | null;
-  rawClustersBuffer: Float32Array | null;
+  rawClustersBuffer: Uint16Array | null;
   myceliumLines: LineSegments | null;
   myceliumGroup: Group | null;
   myceliumCoreLines: LineSegments | null;
